Move archive code to separate pkg
